feat: pluginのVite/Astro入口を分離
