+++ packages/plugin/package.json
@@ -39,6 +39,10 @@
       "types": "./dist/builder/vite.d.ts",
       "import": "./dist/builder/vite.js"
     },
+    "./astro": {
+      "types": "./dist/builder/astro.d.ts",
+      "import": "./dist/builder/astro.js"
+    },
     "./builder": {
       "types": "./dist/builder/index.d.ts",
       "import": "./dist/builder/index.js"
