+++ packages/plugin/src/index.ts
@@ -1,2 +1,1 @@
-export { default, lismCss, lismCssAstro } from './builder/vite';
-export type { LismCssOptions } from './builder/vite';
+export type { LismCssOptions } from './builder/shared';
