+++ templates/blog/astro/techlog/astro.config.mjs
@@ -5,7 +5,7 @@ import expressiveCode from 'astro-expressive-code';
 import remarkDirective from 'remark-directive';
 import { rehypeHeadingIds } from '@astrojs/markdown-remark';
 import rehypeAutolinkHeadings from 'rehype-autolink-headings';
-import { lismCssAstro } from '@lism-css/plugin';
+import { lismCss } from '@lism-css/plugin/astro';
 import { remarkDirectiveHandler } from './src/lib/remark-directive.mjs';
 import { remarkLinkCard } from './src/lib/remark-link-card.mjs';
 import { remarkWikiLink } from './src/lib/remark-wiki-link.mjs';
@@ -44,7 +44,7 @@ export default defineConfig({
         return item;
       },
     }),
-    lismCssAstro({ purge: { report: true } }),
+    lismCss({ purge: { report: true } }),
   ],
   markdown: {
     // :::note などの directive 記法を <Callout> に変換
